+++ src/stores/callStore2.ts
@@ -491,10 +491,14 @@ export class CallStore {
     if (Platform.OS !== 'web') {
       uuid = newUuid().toUpperCase()
       this.callkeepUuidPending = uuid
-      if (Platform.OS == 'android') {
+      if (Platform.OS === 'android') {
         RNCallKeep.startCall(uuid, 'Brekeke phone', number)
       } else {
         RNCallKeep.startCall(uuid, number, number, 'generic', false)
+        // ios if sip call get response INVITE 18x quickly in 50ms - 130ms
+        // add time out to make sure audio active (didDeactivateAudioSession)
+        // before sip call established
+        await waitTimeout(1000)
       }
       this.setAutoEndCallKeepTimer(uuid)
     }