+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 #### 2.14.3
 
 - Fix it should display well with html characters < >
+- Fix it should play RBT correctly with 18x status code (issue 864)
+- Fix ios it should have call audio correctly with low latency response (issue 861, 863)
 - Fix it should not display avatar twice in case PhoneAppli enabled (issue 860)
 - Fix it should navigate to PhoneAppli if enabled and tapping on a missed call notification (issue 856)
 - Fix android it should handle the case user reject permission default dialer phone app (issue 855)